Revert "update sqlcipher"
This reverts commit 9dd5e3ac909e2b140574ac4cd849b33a0585fa4c.

because of: https://github.com/sqlcipher/sqlcipher/issues/582

SUMMARY: ThreadSanitizer: lock-order-inversion (potential deadlock) sqlcipher/sqlite3.c:40233 in unixEnterMutex

problem was reported upstream to SQLite as well. The determination by the SQLite team was that "the deadlock is not reachable due to other constraints within the overall system." Even so, an upstream fix has already be committed that resolves the problem on SQLite's trunk. However, it has not been included in an SQLite release yet. Once it is, we will pull the fix into SQLCipher. I will leave this issue open until that time.